requirements.txt
coverage==4.5.4
Django==2.2.4
django-cors-headers==3.1.0
django-rest-framework==0.1.0
djangorestframework==3.10.2
djangorestframework-api-key==1.4.0
psycopg2-binary==2.8.3
python-dotenv==0.10.3
pytz==2019.2
sqlparse==0.3.0
